+++ src/main/java/io/fjsn/chirp/internal/schema/FieldSchema.java
@@ -1,4 +1,4 @@
-package io.fjsn.chirp.internal;
+package io.fjsn.chirp.internal.schema;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Type;
